+++ src/maps/babel/plugins.ts
@@ -10,6 +10,18 @@ const plugins: Map = {
   "@babel/plugin-proposal-private-methods": {
     nameInRegistry: "@babel/plugin-proposal-private-methods",
   },
+  "@babel/plugin-transform-runtime": {
+    nameInRegistry: "@babel/plugin-transform-runtime",
+  },
+  "@babel/plugin-proposal-class-properties": {
+    nameInRegistry: "@babel/plugin-proposal-class-properties",
+  },
+  "@babel/plugin-proposal-export-namespace-from": {
+    nameInRegistry: "@babel/plugin-proposal-export-namespace-from",
+  },
+  "styled-components": {
+    nameInRegistry: "babel-plugin-styled-components",
+  },
 };
 
 export default plugins;