+++ src/maps/eslint/extends.ts
@@ -19,6 +19,21 @@ const extendsMap: Map = {
   "plugin:@typescript-eslint/recommended": {
     nameInRegistry: "@typescript-eslint/eslint-plugin",
   },
+  airbnb: {
+    nameInRegistry: "eslint-config-airbnb",
+  },
+  prettier: {
+    nameInRegistry: "eslint-config-prettier",
+  },
+  "prettier/react": {
+    nameInRegistry: "eslint-config-prettier",
+  },
+  "plugin:prettier/recommended": {
+    nameInRegistry: "eslint-plugin-prettier",
+  },
+  "plugin:flowtype/recommended": {
+    nameInRegistry: "eslint-plugin-flowtype",
+  },
 };
 
 export default extendsMap;