+++ src/maps/eslint/parser.ts
@@ -4,6 +4,9 @@ const parser: Map = {
   "@typescript-eslint/parser": {
     nameInRegistry: "@typescript-eslint/parser",
   },
+  "babel-eslint": {
+    nameInRegistry: "babel-eslint",
+  },
 };
 
 export default parser;