+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # main (unreleased)
 
+# 0.2.0
+- Add support for Babel plugins: `@babel/plugin-transform-runtime`, `@babel/plugin-proposal-class-properties`, `@babel/plugin-proposal-export-namespace-from`, `styled-components`
+- Add support for Babel presets: `@babel/preset-flow`
+- Add support for ESLint extends: `airbnb`, `prettier`, `prettier/react`, `plugin:prettier/recommended`, `plugin:flowtype/recommended`
+- Add support for ESLint plugins: `flowtype`, `@typescript-eslint`, `sonarjs`
+- Add support for ESLint parser: `babel-eslint`
+
 # 0.1.2
 - Extend and improve usage, API and known issues sections in the README.md.